pyproject.toml
[project]
name = "auto-recon-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Rocha",email = "lecoverde10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.1.0,<8.0.0)",
    "rq (>=2.6.1,<3.0.0)",
    "idna (>=3.11,<4.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.10"
testcontainers = "^4.12.0"
pytest-asyncio = "^1.1.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.5"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.lint.pylint]
max-nested-blocks = 10
max-statements = 85
max-branches = 25
max-locals = 30

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'uvicorn auto_recon_api.app:app --host 0.0.0.0'
pre_test = 'task lint'
test = 'pytest -s -x --cov=auto_recon_api -vv'
post_test = 'coverage html'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]